+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
-- New `drawer-open-trigger` interface.
+- New `drawer-trigger` interface.
 
 ## [0.7.0] - 2020-02-28
 ### Changed